requirements.txt
streamlit==1.32.0
youtube-transcript-api==0.6.2
google-generativeai==0.3.2
google-cloud-speech==2.24.1
google-cloud-storage==2.14.0
pytube==15.0.0
yt-dlp==2024.3.10 